+++ src/main/java/io/dpopkov/knowthenix/services/BaseService.java
@@ -1,0 +1,18 @@
+package io.dpopkov.knowthenix.services;
+
+import io.dpopkov.knowthenix.services.dto.BaseDto;
+
+import java.util.List;
+
+public interface BaseService <T extends BaseDto> {
+
+    T create(T dto);
+
+    T getById(Long id);
+
+    List<T> getAll();
+
+    T update(T dto);
+
+    void delete(Long id);
+}