+++ src/main/java/io/dpopkov/knowthenix/services/CategoryService.java
@@ -1,0 +1,6 @@
+package io.dpopkov.knowthenix.services;
+
+import io.dpopkov.knowthenix.services.dto.CategoryDto;
+
+public interface CategoryService extends BaseService<CategoryDto> {
+}