+++ src/main/java/io/dpopkov/knowthenix/services/CategoryServiceException.java
@@ -1,0 +1,7 @@
+package io.dpopkov.knowthenix.services;
+
+public class CategoryServiceException extends RuntimeException {
+    public CategoryServiceException(String message) {
+        super(message);
+    }
+}