+++ src/main/java/io/dpopkov/knowthenix/services/dto/BaseDto.java
@@ -1,0 +1,10 @@
+package io.dpopkov.knowthenix.services.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class BaseDto {
+    private Long id;
+}