+++ src/main/java/io/dpopkov/knowthenix/services/dto/CategoryDto.java
@@ -1,0 +1,15 @@
+package io.dpopkov.knowthenix.services.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class CategoryDto extends BaseDto {
+    private String name;
+    private String description;
+}