+++ history.md
@@ -2,4 +2,5 @@
 
 (in reversed order)
 
+* CRUD for Category: service.
 * CRUD for Category: entity and repository.